KOGITO-9210 Moved test module
