+++ quarkus/addons/knative/eventing/integration-tests/pom.xml
@@ -5,12 +5,12 @@
     <modelVersion>4.0.0</modelVersion>
     <parent>
         <groupId>org.kie.kogito</groupId>
-        <artifactId>kogito-quarkus-integration-tests</artifactId>
+        <artifactId>kogito-addons-quarkus-knative-eventing-parent</artifactId>
         <version>2.0.0-SNAPSHOT</version>
     </parent>
 
-    <artifactId>integration-tests-quarkus-knative-eventing</artifactId>
-    <name>Kogito :: Integration Tests :: Quarkus :: Knative :: Eventing</name>
+    <artifactId>kogito-addons-quarkus-knative-eventing-integration-tests</artifactId>
+    <name>Kogito Add-On Knative Eventing - Integration Tests</name>
 
     <dependencyManagement>
         <dependencies>
