+++ quarkus/addons/knative/eventing/integration-tests/src/main/resources/hello.sw.json
