+++ quarkus/addons/knative/eventing/integration-tests/src/test/java/org/kie/kogito/addons/quarkus/knative/eventing/KSinkInjectionHealthCheckDisabledProfile.java
@@ -13,25 +13,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kie.kogito.integrationtests.quarkus.knative.eventing;
+package org.kie.kogito.addons.quarkus.knative.eventing;
 
-import javax.inject.Inject;
+import java.util.Map;
 
-import org.eclipse.microprofile.config.Config;
-import org.hamcrest.Matcher;
+import io.quarkus.test.junit.QuarkusTestProfile;
 
-import static io.restassured.RestAssured.given;
+public class KSinkInjectionHealthCheckDisabledProfile implements QuarkusTestProfile {
 
-abstract class AbstractKSinkInjectionHealthCheckTest {
-
-    @Inject
-    Config config;
-
-    protected void assertHealthChecks(Matcher<?> matcher) {
-        given()
-                .when()
-                .get("/q/health/live")
-                .then()
-                .body("checks", matcher);
+    @Override
+    public Map<String, String> getConfigOverrides() {
+        return Map.of(KSinkInjectionHealthCheck.CONFIG_ALIAS, "false");
     }
 }
