+++ quarkus/addons/knative/eventing/pom.xml
@@ -16,6 +16,7 @@
   <modules>
     <module>runtime</module>
     <module>deployment</module>
+    <module>integration-tests</module>
   </modules>
 
 </project>
